# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Cotações API"
  },
  "host": "https://sistemas-cotacao-backend.herokuapp.com",
  "basePath": "/",
...
  "paths": {
    "/usuarios": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Adiciona usuario",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nome"
          },
          {
            "name":"cargo"
          },
          {
            "name":"login"
          },
          {
            "name":"senha"
          }
        ],
        "responses": {
          "200": {
            "description": "Valido"
          },
          "500": {
            "description": "Erro"
          }
        }
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "summary": "Update Usuario",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nome"
          },
          {
            "name":"cargo"
          },
          {
            "name":"login"
          },
          {
            "name":"senha"
          }
        ],
        "responses": {
          "200": {
            "description": "Valido"
          },
          "500": {
            "description": "Erro"
          }
        }
      },
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Pegar todos Usuarios",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          
        ],
        "responses": {
          "200": {
            "description": "Valido"
          },
          "500": {
            "description": "Erro"
          }
        }
      }
    },
    "/usuarios/:id": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Pegar Usuario Especifico",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          
        ],
        "responses": {
          "200": {
            "description": "Valido"
          },
          "500": {
            "description": "Erro"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "summary": "Deletar Usuario Especifico",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          
        ],
        "responses": {
          "200": {
            "description": "Valido"
          },
          "500": {
            "description": "Erro"
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Login Usuario",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "login"
          },
          {
            "name":"senha"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario Autenticado"
          },
          "500": {
            "description": "Usuario ou senha invalido"
          }
        }
      }
    },
...
  }
}